Rename AsyncSyncBox to InitializableState
